+++ include/quill/std/Optional.h
@@ -8,6 +8,7 @@
 #include "quill/core/Attributes.h"
 #include "quill/core/Codec.h"
 #include "quill/core/DynamicFormatArgStore.h"
+#include "quill/core/InlinedVector.h"
 #include "quill/core/Utf8Conv.h"
 
 #include "quill/bundled/fmt/std.h"
@@ -27,7 +28,7 @@ QUILL_BEGIN_NAMESPACE
 template <typename T>
 struct Codec<std::optional<T>>
 {
-  static size_t compute_encoded_size(std::vector<size_t>& conditional_arg_size_cache,
+  static size_t compute_encoded_size(detail::SizeCacheVector& conditional_arg_size_cache,
                                      std::optional<T> const& arg) noexcept
   {
     // We need to store the size of the vector in the buffer, so we reserve space for it.
@@ -42,7 +43,7 @@ struct Codec<std::optional<T>>
     return total_size;
   }
 
-  static void encode(std::byte*& buffer, std::vector<size_t> const& conditional_arg_size_cache,
+  static void encode(std::byte*& buffer, detail::SizeCacheVector const& conditional_arg_size_cache,
                      uint32_t& conditional_arg_size_cache_index, std::optional<T> const& arg) noexcept
   {
     Codec<bool>::encode(buffer, conditional_arg_size_cache, conditional_arg_size_cache_index, arg.has_value());